requirements.txt
# Minimal dependencies for Azure OpenAI client
langchain-openai>=0.1.0
azure-identity>=1.12.0
python-dotenv>=1.0.0
